Extracted Cluster Interface
The KubernetesCluster implementation replaces Kubernetes Deployments and KubernetesPods.
One Kubernetes deployment is one app in the cluster. Each app can have multiple replicas. (ClusterAppReplica)

Additionally the main package was renamed to kpad.
