+++ build.gradle
@@ -45,10 +45,10 @@ license {
     }
 }
 
-group = 'de.qaware.cloud.nativ.k8s'
+group = 'de.qaware.cloud.nativ.kpad'
 version = '1.0.1'
 
-mainClassName = 'de.qaware.cloud.nativ.k8s.KubepadApplication'
+mainClassName = 'de.qaware.cloud.nativ.kpad.KubepadApplication'
 applicationDefaultJvmArgs = ['-Djava.library.path=../lib']
 
 sourceCompatibility = 1.8
@@ -60,6 +60,10 @@ dependencies {
 
     compile 'io.fabric8:kubernetes-api:2.2.126'
     compile 'io.fabric8:kubernetes-client:1.3.95'
+
+    compile 'com.squareup.retrofit2:retrofit:2.1.0'
+    compile 'com.squareup.retrofit2:converter-gson:2.1.0'
+    //compile 'com.github.mohitsoni:marathon-client:0.4.2'
 
     compile ':LeapJava:2.3.1'
