--- src/main/kotlin/de/qaware/cloud/nativ/k8s/launchpad/MidiSystemProducer.kt
@@ -1,51 +1,0 @@
-/*
- * The MIT License (MIT)
- *
- * Copyright (c) 2016 QAware GmbH, Munich, Germany
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- */
-package de.qaware.cloud.nativ.k8s.launchpad
-
-import javax.enterprise.context.ApplicationScoped
-import javax.enterprise.inject.Default
-import javax.enterprise.inject.Disposes
-import javax.enterprise.inject.Produces
-import javax.sound.midi.MidiSystem
-import javax.sound.midi.Receiver
-import javax.sound.midi.Transmitter
-
-/**
- * Global CDI producer for any MIDI related instances.
- */
-@ApplicationScoped
-class MidiSystemProducer {
-
-    @Produces
-    @Default
-    fun transmitter(): Transmitter = MidiSystem.getTransmitter()
-
-    @Produces
-    @Default
-    fun receiver(): Receiver = MidiSystem.getReceiver()
-
-    fun close(@Disposes transmitter: Transmitter) = transmitter.close()
-
-    fun close(@Disposes receiver: Receiver) = receiver.close()
-}
