+++ src/main/kotlin/de/qaware/cloud/nativ/kpad/ClusterAppReplica.kt
@@ -1,0 +1,6 @@
+package de.qaware.cloud.nativ.kpad
+
+interface ClusterAppReplica {
+    fun phase() : ClusterNode.Phase
+    fun name() : String
+}
