+++ src/main/kotlin/de/qaware/cloud/nativ/kpad/ExecutorProducer.kt
@@ -1,0 +1,50 @@
+/*
+ * The MIT License (MIT)
+ *
+ * Copyright (c) 2016 QAware GmbH, Munich, Germany
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+package de.qaware.cloud.nativ.kpad
+
+import java.util.concurrent.ExecutorService
+import java.util.concurrent.Executors
+import java.util.concurrent.ScheduledExecutorService
+import javax.enterprise.context.ApplicationScoped
+import javax.enterprise.inject.Any
+import javax.enterprise.inject.Disposes
+import javax.enterprise.inject.Produces
+import javax.inject.Named
+
+/**
+ * A CDI producer bean for different executor service instances.
+ */
+@ApplicationScoped
+class ExecutorProducer {
+
+    @Produces
+    @Named("default")
+    fun executor(): ExecutorService = Executors.newFixedThreadPool(2)
+
+    @Produces
+    @Named("scheduled")
+    fun scheduled(): ScheduledExecutorService = Executors.newScheduledThreadPool(2)
+
+    fun shutdown(@Disposes @Any executorService: ExecutorService) = executorService.shutdown()
+}
